requirements.txt
django
djangorestframework
channels
openai
daphne
python-dotenv
